wip: add json part file
